Move text around
I am tyring to orgainize each section of the notes in the oreder they will be presented.  The goal is to get the workshop writing code in intervals as we work through the notes.  A secondary goal is to have the frist coding session as elary as is reasonable so I don't end up lecturing for 3 hours at the start.