Implementation of Task 2
A function was written to iextraction the data from task two.
Task two asks the user to pick the eyes of cats in the image.
This results in a list of x and y coords for each eye.
The code loops through each entry in the data - by class, user, and
subject id.
The code then loops through all pairs of coordinates and saves them
into a list of tuples.
This is then concatenated and returned with the class, user, and
subject id.